libc: Move ofdtx interfaces to ofdtx.c
